# openapi.yaml
openapi: 3.0.3
info:
  title: 焚信 BurnMsg API
  version: 2.0.0
  description: |-
    焚信（BurnMsg）是一款企业内部加密即时通讯（IM）应用，支持云聊天与秘密聊天、阅后即焚消息。本文件为其后端（NestJS + TypeORM + MySQL）的标准 OpenAPI 3.0.3 接口文档。

    【账号体系】
    - 不开放自助注册，所有账号由管理后台统一开通（POST /accounts，仅 admin）。
    - 用户名即公司内部手机号（1 开头 11 位数字）。
    - 新开通 / 被重置密码的账号 force_change_pwd=true，前端应强制跳转修改密码，修改成功后该标记置为 false。

    【认证方式】
    - 受保护接口在请求头携带：Authorization: Bearer <access_token>。
    - 双 token 机制：access_token（JWT_SECRET 签名，默认有效期 7200 秒）+ refresh_token（独立的 JWT_REFRESH_SECRET 签名，默认有效期 2592000 秒 / 30 天，payload 带 type=refresh 标记防混用）。
    - access_token 过期后用 refresh_token 调 POST /auth/refresh-token 换取新 token 对。
    - 即时吊销：JWT 守卫除验签外，每个受保护请求都会实时查库校验「账号仍为 active」且「签发 token 的设备记录仍存在」；账号被停用或设备被下线后，已签发的 token 立即失效（全端即时强制下线）。

    【统一响应格式】
    - 成功：{ "code": 0, "message": "ok 或各接口自定义提示语", "data": <业务数据，无内容时为 null> }
    - 失败：{ "code": <HTTP 状态码>, "message": "<错误描述；参数校验失败时为多条错误以 '; ' 拼接>" }

    【通用约定】
    - 分页参数：page（默认 1）、pageSize（默认 20，最大 200；通讯录列表默认 50）。
    - 请求中的时间字段使用 ISO8601 字符串；响应中的时间为服务端 datetime。
    - 所有业务 ID 均为 UUID v4 字符串。
    - 角色：user / admin。标注「仅管理员」的接口仅 admin 可调用，普通用户返回 403。
    - 强制改密拦截：force_change_pwd=true 的账号（新开通 / 被重置密码后），除 POST /auth/change-password 与 GET /auth/profile 外，调用其他任何受保护接口一律返回 403（FORCE_CHANGE_PASSWORD），前端应引导其先完成改密再进入主界面。
    - 软删除账号：被 DELETE /accounts/{id} 软删的账号立即吊销全部 token 并强制下线，其历史消息保留、前端统一展示「已注销用户」；账号列表默认不含已软删账号，加 show_deleted=1 可查看。
servers:
  - url: http://localhost:9091/api/v1
...
security:
  - bearerAuth: []
...
paths:
...
  /messages:
    post:
      tags: [消息]
      summary: 发送消息（文字/图片/语音/视频/文件），支持回复引用与阅后即焚
      description: |-
        【用途】向指定会话发送一条消息，支持文本及图片/语音/视频/文件附件，支持回复引用与阅后即焚。

        【调用时机】聊天页发送消息时；附件类消息需先调 POST /upload 拿到 url 再调本接口。

        【业务规则与注意事项】
        - 仅会话成员可向该会话发消息，非成员返回 403（修复了旧版越权发消息漏洞）。
        - type=text 时 content 必填（400 文本消息 content 不能为空）；type 为 image/voice/video/file 时 file_url 必填（400 xxx 消息 file_url 不能为空）。
        - 阅后即焚：传 destroy_at（ISO8601 绝对时间，如 5 分钟后的时间点），到期由服务端定时任务（每分钟扫描一批，每批最多 500 条）自动销毁——删除该消息全部回执、清空 content 与 file_url、置 is_destroyed=true；消息列表接口另有「到期即不返回」的过滤兜底（双保险）。源码未限定销毁时长的离散枚举，由客户端自行计算目标时间传入。
        - 发送成功后：会话 last_message_at 更新为当前时间（影响会话列表排序）；为除发送者外的每位成员各创建一条消息回执（is_delivered=false, is_read=false）。
        - reply_to_id 传入被回复消息的 Message.id 实现引用回复。
        - E2E 加密：cipher_text / cipher_nonce / sender_ephemeral_pubkey 三个字段「全有或全无」——全传则按加密消息入库（is_encrypted=true，content 可置空或留占位），只传部分返回 400；都不传则为普通明文消息。协商流程见 docs/E2E_ENCRYPTION.md。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversation_id, type]
              properties:
                conversation_id:
                  type: string
                  format: uuid
                  description: 目标会话 ID（UUID）
                  example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
                type:
                  type: string
                  enum: [text, image, voice, video, file]
                  description: 消息类型：text 文本 / image 图片 / voice 语音 / video 视频 / file 文件
                  example: text
                content:
                  type: string
                  description: 文本内容；type=text 时必填
                  example: 会议纪要已发，请查收
                file_url:
                  type: string
                  maxLength: 500
                  description: 附件 URL（POST /upload 返回的 url）；非 text 类型必填，最长 500 字符
                  example: /uploads/9b1d2c3e-4f5a-4b6c-8d7e-9f0a1b2c3d4e.pdf
                file_name:
                  type: string
                  maxLength: 200
                  description: 附件原始文件名，最长 200 字符
                  example: 会议纪要.pdf
                file_size:
                  type: integer
                  description: 附件大小（字节）
                  example: 204800
                reply_to_id:
                  type: string
                  format: uuid
                  description: 被回复消息的 ID（可选，实现引用回复）
                  example: 5e6f7a8b-9c0d-41e2-83f4-a5b6c7d8e9f0
                destroy_at:
                  type: string
                  format: date-time
                  description: 阅后即焚销毁时间（ISO8601 绝对时间）；到期后服务端定时任务自动销毁该消息；不传则永久保留
                  example: '2026-08-26T12:05:00.000Z'
                cipher_text:
                  type: string
                  description: E2E 加密密文（AES-256-GCM，base64）；明文消息不传
                cipher_nonce:
                  type: string
                  pattern: '^[A-Za-z0-9+/=_-]{12,24}$'
                  description: GCM nonce（base64）；与 cipher_text 配套
                sender_ephemeral_pubkey:
                  type: string
                  pattern: '^[A-Za-z0-9+/_-]{43,44}$'
                  description: 发送方临时 X25519 公钥（base64，43-44 字符），接收方用它做 ECDH 还原密钥
      responses:
        '200':
          description: 发送成功（message 为「发送成功」），data 为已入库的 Message
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 发送成功 }
                  data:
                    $ref: '#/components/schemas/Message'
        '400':
          description: 参数校验失败；或 text 消息缺 content；或非 text 消息缺 file_url
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非会话成员禁止向该会话发消息（Access denied to this conversation）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /messages/{conversationId}:
    get:
      tags: [消息]
      summary: 会话消息历史（倒序拉取，已到期销毁的消息被过滤）
      description: |-
        【用途】拉取指定会话的历史消息。

        【调用时机】进入聊天页首次加载（不带 before，拉最新一页）；向上滚动加载更早消息（带 before 游标翻页）。

        【业务规则与注意事项】
        - 仅会话成员可访问，非成员返回 403。
        - 默认返回最新 limit 条（默认 50）；向上翻页时把当前已加载的最旧一条消息的 created_at 作为 before 传入，服务端返回 created_at 早于该时间的消息，实现历史消息游标分页。
        - 返回数组按时间正序（旧→新）排列，便于直接渲染。
        - 已销毁（is_destroyed=true）及已到销毁时间（destroy_at <= 当前时间）的消息不返回，即使定时销毁任务尚未执行（过滤兜底）。
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 会话 ID（UUID）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
        - name: before
          in: query
          schema: { type: string, format: date-time }
          description: 分页游标：传当前已加载的最旧一条消息的 created_at（ISO8601），返回早于该时间的消息；不传则拉最新一页
          example: '2026-08-26T08:30:00.000Z'
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
          description: 每页条数，默认 50
          example: 50
      responses:
        '200':
          description: 成功，data 为 Message 数组（按时间正序）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非会话成员（Access denied to this conversation）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /messages/{id}:
    put:
      tags: [消息]
      summary: 编辑消息（置 is_edited=true）
      description: |-
        【用途】编辑自己已发送的消息内容。

        【调用时机】用户在聊天页对自己发出的消息选择「编辑」时。

        【业务规则与注意事项】
        - 仅消息发送者可编辑，非发送者返回 403（You can only edit your own messages）。
        - 时间窗：仅发送后 48 小时内可编辑，超过返回 400（Cannot edit messages older than 48 hours）。
        - 已撤回的消息不可再编辑，返回 400（Cannot edit a recalled message）。
        - 编辑成功后 is_edited=true，前端应展示「已编辑」标记。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 消息 ID（UUID）
          example: 5e6f7a8b-9c0d-41e2-83f4-a5b6c7d8e9f0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: 新的消息内容（必填）
                  example: 会议纪要已更新，请查收
      responses:
        '200':
          description: 编辑成功（message 为「消息已编辑」），data 为更新后的 Message
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 消息已编辑 }
                  data:
                    $ref: '#/components/schemas/Message'
        '400':
          description: content 为空；消息已撤回不可编辑；消息发送超过 48 小时不可编辑
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非消息发送者（You can only edit your own messages）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 消息不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /messages/{id}/recall:
    post:
      tags: [消息]
      summary: 撤回消息（置 is_recalled=true）
      description: |-
        【用途】撤回自己已发送的消息。

        【调用时机】用户在聊天页对自己发出的消息选择「撤回」时。

        【业务规则与注意事项】
        - 仅消息发送者可撤回，非发送者返回 403（You can only recall your own messages）。
        - 时间窗：仅发送后 48 小时内可撤回，超过返回 400（Cannot recall messages older than 48 hours）。
        - 撤回为软删除：仅置 is_recalled=true，消息记录与内容仍保留在库中，前端应展示「已撤回」占位。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 消息 ID（UUID）
          example: 5e6f7a8b-9c0d-41e2-83f4-a5b6c7d8e9f0
      responses:
        '200':
          description: 撤回成功（message 为「消息已撤回」，data 为 null）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 消息已撤回 }
                  data: { type: object, nullable: true, example: null }
        '400':
          description: 消息发送超过 48 小时不可撤回
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非消息发送者（You can only recall your own messages）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 消息不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /messages/{id}/read:
    post:
      tags: [消息]
      summary: 将会话标记为已读（单条 UPDATE，同步本人全部回执）
      description: |-
        【用途】把指定会话标记为已读。

        【调用时机】用户打开某个聊天会话、浏览到最新消息时。

        【业务规则与注意事项】
        - 注意：路径参数 id 为「会话 ID」（与旧版 API 保持一致），不是消息 ID。
        - 仅会话成员可操作，非成员返回 403。
        - 服务端用一条 UPDATE（含子查询）把本人在该会话下全部未读回执置为 is_read=true、read_at=当前时间；并把本人成员记录的 last_read_message_id 更新为该会话最新一条消息。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 会话 ID（UUID；注意此处不是消息 ID）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
      responses:
        '200':
          description: 标记成功（message 为「已读标记成功」，data 为 null）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 已读标记成功 }
                  data: { type: object, nullable: true, example: null }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非会话成员（Access denied to this conversation）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /messages/{id}/receipt:
    get:
      tags: [消息]
      summary: 消息回执列表（送达/已读明细，含用户姓名）
      description: |-
        【用途】查看一条消息的逐人回执明细（是否送达、是否已读、已读时间）。

        【调用时机】群聊中点击自己消息的「已读/未读」详情时。

        【业务规则与注意事项】
        - 仅该消息所在会话的成员可查看，非成员返回 403（修复了旧版任意用户可查漏洞）。
        - 消息不存在返回 404。
        - 回执在消息发送时为除发送者外的每位成员各创建一条；每条附 user_display_name 便于直接展示。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 消息 ID（UUID）
          example: 5e6f7a8b-9c0d-41e2-83f4-a5b6c7d8e9f0
      responses:
        '200':
          description: 成功，data 为回执数组（含用户姓名）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MessageReceipt'
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非会话成员（Access denied to this conversation）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 消息不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        JWT access token。登录（POST /auth/login）或刷新（POST /auth/refresh-token）后获得，放入请求头：
        Authorization: Bearer <access_token>
        默认有效期 7200 秒；每个受保护请求都会实时校验账号状态与设备存在性，账号停用或设备下线后 token 立即失效。
  schemas:
    Error:
      type: object
      description: 统一错误响应（code 为 HTTP 状态码；参数校验失败时 message 为多条错误以 '; ' 拼接）
      properties:
        code:
          type: integer
          description: HTTP 状态码（400/401/403/404/409/413/429/500）
          example: 400
        message:
          type: string
          description: 错误描述
          example: 手机号格式不正确
      required: [code, message]
...
    Message:
      type: object
      description: 消息对象（阅后即焚消息到期后 content/file_url 被清空且 is_destroyed=true）
      properties:
        id:
          type: string
          format: uuid
          description: 消息 ID（UUID v4）
          example: 5e6f7a8b-9c0d-41e2-83f4-a5b6c7d8e9f0
        conversation_id:
          type: string
          format: uuid
          description: 所属会话 ID
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
        sender_id:
          type: string
          format: uuid
          description: 发送者用户 ID
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        type:
          type: string
          enum: [text, image, voice, video, file]
          description: 消息类型：text 文本 / image 图片 / voice 语音 / video 视频 / file 文件
          example: text
        content:
          type: string
          nullable: true
          description: 文本内容；附件类消息或被销毁后为 null
          example: 会议纪要已发，请查收
        file_url:
          type: string
          nullable: true
          description: 附件 URL（POST /upload 返回）；文本消息或被销毁后为 null
          example: /uploads/9b1d2c3e-4f5a-4b6c-8d7e-9f0a1b2c3d4e.pdf
        file_name:
          type: string
          nullable: true
          description: 附件原始文件名
          example: 会议纪要.pdf
        file_size:
          type: integer
          nullable: true
          description: 附件大小（字节）
          example: 204800
        reply_to_id:
          type: string
          format: uuid
          nullable: true
          description: 被回复消息的 ID（引用回复），无则 null
          example: null
        is_edited:
          type: boolean
          description: 是否被编辑过（编辑成功后为 true，前端展示「已编辑」）
          example: false
        is_recalled:
          type: boolean
          description: 是否已撤回（撤回为软删除，前端展示「已撤回」占位）
          example: false
        is_destroyed:
          type: boolean
          description: 阅后即焚是否已销毁（销毁后 content/file_url 已清空）
          example: false
        destroy_at:
          type: string
          format: date-time
          nullable: true
          description: 阅后即焚销毁时间，到期服务端自动销毁；非焚毁消息为 null
          example: null
        is_encrypted:
          type: boolean
          description: 是否 E2E 端到端加密消息；为 true 时正文在 cipher_text，content 为占位/空
          example: false
        cipher_text:
          type: string
          nullable: true
          description: E2E 密文（AES-256-GCM，base64）；明文消息为 null
          example: null
        cipher_nonce:
          type: string
          nullable: true
          description: GCM nonce（base64）；明文消息为 null
          example: null
        sender_ephemeral_pubkey:
          type: string
          nullable: true
          description: 发送方临时 X25519 公钥（base64）；明文消息为 null
          example: null
        created_at:
          type: string
          format: date-time
          description: 发送时间（历史消息分页 before 游标传此值）
          example: '2026-08-26T08:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: '2026-08-26T08:00:00.000Z'
...
    MessageReceipt:
      type: object
      description: 消息回执（发送时为除发送者外的每位成员各创建一条）
      properties:
        id:
          type: string
          format: uuid
          description: 回执记录主键
          example: 8b9c0d1e-2f3a-44b5-96c7-d8e9f0a1b2c3
        message_id:
          type: string
          format: uuid
          description: 消息 ID
          example: 5e6f7a8b-9c0d-41e2-83f4-a5b6c7d8e9f0
        user_id:
          type: string
          format: uuid
          description: 接收者用户 ID
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        is_delivered:
          type: boolean
          description: 是否已送达
          example: false
        is_read:
          type: boolean
          description: 是否已读
          example: true
        read_at:
          type: string
          format: date-time
          nullable: true
          description: 已读时间，未读为 null
          example: '2026-08-26T08:05:00.000Z'
        created_at:
          type: string
          format: date-time
          description: 回执创建时间（即消息发送时间）
          example: '2026-08-26T08:00:00.000Z'
        user_display_name:
          type: string
          description: 接收者显示名（用户不存在时为 Unknown）
          example: 李四
...